src/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic-settings==2.9.1
aiofiles==24.1.0
langchain-community==0.3.25
PyMuPDF==1.26.1
langchain-text-splitters==0.3.8
motor==3.7.1
openai==1.91.0
cohere== 5.15.0
qdrant-client== 1.14.3
SQLAlchemy==2.0.41
asyncpg==0.30.0
alembic==1.16.4
psycopg2==2.9.10
pgvector==0.4.1
nltk==3.9.1

# Monitoring and Metrics 
prometheus-client==0.22.1
starlette-exporter==0.23.0

# Health Checks
fastapi-health==0.4.0
